Whole lot of stuff.
 * Added Metadata to resource allowing storing of data by systems.
 * Moved asset parsing of the io thread.
 * Work on ogl implementation.
 * Fixed some bugs.
 * Experimented with implementions for the defered game loop.

+++ tungsten/src/asset/data.rs
@@ -233,7 +233,7 @@ pub static SPHERE_NORMALS: [f32; 342] = [
                 -8.8501998e-18,-1.00000000,-6.2393909e-17,
             ];
 
-pub static SPHERE_INDECIES: [usize; 672] = [
+pub static SPHERE_INDECIES: [u32; 672] = [
                 1,18,17,
                 1,32,16,
                 1,113,2,
